moved files to the right directories
