+++ src/components/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import './App.css';
-import BooksForm from './components/BookForm';
-import BooksList from './components/BookList';
+import '../App.css';
+import BooksForm from './BookForm';
+import BooksList from '../containers/BookList';
 
 const App = () => (
   <div className="App">
