+++ src/index.js
@@ -4,7 +4,7 @@ import './index.css';
 import { createStore } from 'redux';
 import { Provider } from 'react-redux';
 import rootReducer from './reducers';
-import App from './App';
+import App from './components/App';
 import * as serviceWorker from './serviceWorker';
 
 const store = createStore(rootReducer);
